[refactor] added delete garden and its plants to myGardenController.js

+++ client/app/myGarden/myGardenController.js
@@ -13,7 +13,9 @@ myGarden.controller('myGardenController', ['Plants', '$state', '$window',  funct
   that.count = 0;
   that.resultPlants;
   that.lists;
-  
+  that.gardenArray = [];
+
+
   that.dropCallback = function(event, index, item, external, type){
     var plant = {plantId: item.plantId};
     var garden = {gardenName: item.bucket};
@@ -39,7 +41,7 @@ myGarden.controller('myGardenController', ['Plants', '$state', '$window',  funct
       .then(function(results) {
         for(var i = 0; i < results.length; i++){
           that.gardens[results[i].id] = results[i].gardenName
-        } 
+        }
         that.lists = that.setList(that.resultPlants)
         that.formatGardenForSandbox();
       })
@@ -68,31 +70,53 @@ myGarden.controller('myGardenController', ['Plants', '$state', '$window',  funct
       .catch(function(error) {
         console.log(error);
       });
-  }; 
+  };
 
   that.deleteGarden = function(){
-    if(that.data.gardenDelete){
-      Plants.deleteGarden(that.data)
-      .then(function(results) {
-        console.log(results, 'RESULTS IN DELETE GARDEN CONTROLLER');
-      })
-      .catch(function(error){
-        console.log(error, 'ERROR IN DELETE GARDEN CONTROLLER');
-      })
+    if(confirm('Are You sure you want to delete this garden and all its plants?')){
+      if(that.data.gardenDelete){
+        var gardenObj= {};
+        gardenObj.gardenName = that.data.gardenDelete
+        Plants.getGardenPlants(gardenObj)
+          .then(function(results){
+            console.log(results, 'RESULTS IN GETGARDENPLANTSDELETEGAREDJFKLN');
+            for(var i = 0; i<results.length; i++){
+              console.log(results[i].nickname, 'BITCHES GET STICHES')
+              var temp = {};
+              temp.plantDelete = results[i].nickname;
+              console.log(temp, "jlkajfsalkjklewiouriowne")
+              Plants.deletePlant(temp);
+              that.getUserPlants();
+            }
+            Plants.deleteGarden(that.data)
+            .then(function(results) {
+              console.log(results, 'RESULTS IN DELETE GARDEN CONTROLLER');
+              that.getUsersGardens();
+
+            })
+            .catch(function(error){
+              console.log(error, 'ERROR IN DELETE GARDEN CONTROLLER');
+            })
+          })
+          .catch(function(error){
+            console.log(error, 'ERROR IN DELETING PLANTS OF A GARDEN CONTROLLER');
+          })
+
+      }
     }
   };
 
   that.setList = function(arr) {
     var res = arr.reduce(function(obj, cur, i, array) {
       //if theres no garden id or that gardens been seen already, continue
-      if (cur.idOfGarden === '' || obj[cur.idOfGarden]) { return obj } 
-        obj[cur.idOfGarden] = { 
+      if (cur.idOfGarden === '' || obj[cur.idOfGarden]) { return obj }
+        obj[cur.idOfGarden] = {
           label: that.gardens[cur.idOfGarden],
           plants: []
         }
         return obj
         //object to reduce to starts with area for unplanted plants
-      }, { 0: { label: "To Plant!", plants: [] }} 
+      }, { 0: { label: "The Nursery", plants: [] }}
     );
     return res;
   };
